Move more demo related tests
